test/poe2/mpitso/TaskClassTest.java

package poe2.mpitso;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

public class TaskClassTest {
    
    public TaskClassTest() {
    }
    
    @BeforeAll
    public static void setUpClass() {
    }
    
    @AfterAll
    public static void tearDownClass() {
    }
    
    @BeforeEach
    public void setUp() {
    }
    
    @AfterEach
    public void tearDown() {
    }

    /**
     * Test of checkTaskDescription method, of class TaskClass.
     */
    @Test
    public void testCheckTaskDescription() {
        System.out.println("checkTaskDescription");
        TaskClass instance = new TaskClass();
        boolean expResult = false;
        boolean result = instance.checkTaskDescription();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of createTaskID method, of class TaskClass.
     */
    @Test
    public void testCreateTaskID() {
        System.out.println("createTaskID");
        TaskClass instance = new TaskClass();
        String expResult = "";
        String result = instance.createTaskID();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of printTaskDetails method, of class TaskClass.
     */
    @Test
    public void testPrintTaskDetails() {
        System.out.println("printTaskDetails");
        TaskClass instance = new TaskClass();
        String expResult = "";
        String result = instance.printTaskDetails();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of returnTotalHours method, of class TaskClass.
     */
    @Test
    public void testReturnTotalHours() {
        System.out.println("returnTotalHours");
        TaskClass instance = new TaskClass();
        int expResult = 0;
        int result = instance.returnTotalHours();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
